deleting files - file up and download complete

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const Product = require('../models/product');
+const fileHelper = require('../util/file');
 const { validationResult } = require('express-validator');
 
 exports.getAddProduct = (req, res) => {
@@ -135,6 +136,7 @@ exports.postEditProduct = (req, res, next) => {
             product.price = updatedPrice;
             product.description = updatedDesc;
             if (image) {
+                fileHelper.deleteFile(product.imageUrl);
                 product.imageUrl = image.path;
             }
             return product
@@ -153,9 +155,17 @@ exports.postEditProduct = (req, res, next) => {
 exports.deleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
     Product
-        .deleteOne({
-            _id: prodId,
-            userId: req.user._id
+        .findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product
+                .deleteOne({
+                    _id: prodId,
+                    userId: req.user._id
+                });
         })
         .then(() => {
             res.redirect('/admin/products');
